+++ src/Generator/Generator.php
@@ -242,6 +242,15 @@ class Generator
             }
         }
 
+        $requires = [
+            "php"                            => ">=__MIN_PHP_VERSION__",
+            "srag/activerecordconfig"        => ">=0.1.0",
+            "srag/custominputguis"           => ">=0.1.0",
+            "srag/dic"                       => ">=0.1.0",
+            "srag/librariesnamespacechanger" => ">=0.1.0",
+            "srag/removeplugindataconfirm"   => ">=0.1.0"
+        ];
+
         $ilias_plugin = [
             "id"                => "__PLUGIN_ID__",
             "name"              => "__PLUGIN_NAME__",
@@ -269,14 +278,21 @@ class Generator
         ];
 
         if ($this->options->isEnableDevTools()) {
-            $config_ctrl_class[] = "@ilCtrl_isCalledBy srag\DIC\__PLUGIN_NAME__\DevTools\DevToolsCtrl: il__PLUGIN_NAME__ConfigGUI";
+            $requires["srag/devtools"] = ">=0.1.0";
+            $config_ctrl_class[] = "@ilCtrl_isCalledBy srag\DevTools\__PLUGIN_NAME__\DevToolsCtrl: il__PLUGIN_NAME__ConfigGUI";
             $config_ctrl_execute[] = "strtolower(DevToolsCtrl::class):
                 self::dic()->ctrl()->forwardCommand(new DevToolsCtrl(\$this, self::plugin()))";
             $config_tabs[] = "DevToolsCtrl::addTabs(self::plugin())";
             $update_languages[] = "DevToolsCtrl::installLanguages(self::plugin())";
         }
 
+        ksort($requires);
+
         $this->placeholders = [
+            "REQUIRES"                    => implode(",
+    ", array_map(function (string $key, $value) : string {
+                return json_encode($key, JSON_UNESCAPED_SLASHES) . ": " . json_encode($value, JSON_UNESCAPED_SLASHES);
+            }, array_keys($requires), $requires)),
             "AUTHOR_COMMENT"                  => $author_comment,
             "CONFIG_CTRL_CALLS"               => (!empty($config_ctrl_class) ? "
  *