+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Auto generate plugin readme script
 - Footer permanent link
 - Translate plugin SLA to english
+- Adjust new DevTools
 
 ## [2.3.3]
 - Fix